5.1.1 beginnings
changelog:

- changed const floats to doubles
- changed FHitResult.Actor to FHitResult.GetActor()

+++ Source/JaggaJagga/Item.cpp
@@ -295,9 +295,9 @@ void AItem::StartItemCurve(AShooterCharacter* Char)
 		ZCurveTime);
 
 	// Get initial Yaw of the camera
-	const float CameraRotationYaw{ Character->GetFollowCamera()->GetComponentRotation().Yaw };
+	const double CameraRotationYaw{ Character->GetFollowCamera()->GetComponentRotation().Yaw };
 	// Get initial Yaw of the Item
-	const float ItemRotationYaw{ GetActorRotation().Yaw };
+	const double ItemRotationYaw{ (GetActorRotation().Yaw) };
 	// Initial Yaw offset between Camera and Item
 	InterpInitialYawOffset = ItemRotationYaw - CameraRotationYaw;
 }
